Moved async_periodic_task class to a stand-alone utility file.
